+++ Documents/Help/Changelog.md
@@ -4,6 +4,8 @@
 =======
 
 ## 2022-11-xx - Build 2211 - November 2022
+* Added `MostRecentlyUsedFileManager` to the `Krypton.Toolkit.Suite.Extended.IO` module
+* Added 3 new controls in the `Krypton.Toolkit.Suite.Extended.Tool.Strip.Items` module, these are `MRUOpenFileMenuItem`, `MRUSaveAsFileMenuItem` and `MRUSaveFileMenuItem`. Note, these controls only work with text files/documents. If you wish to use the MRU functionality with other file types, please refer to the `MostRecentlyUsedFileManager` in the `Krypton.Toolkit.Suite.Extended.IO` module
 * Implemented [#334](https://github.com/Krypton-Suite/Extended-Toolkit/issues/334), `Krypton.Toolkit.Suite.Extended.Outlook.Grid` - language support German. Thanks to [mikel3361](https://github.com/mikel3361)
 * Implemented [#330](https://github.com/Krypton-Suite/Extended-Toolkit/issues/330), `KryptonProgressBar` add text attribute
 * Resolved [#324](https://github.com/Krypton-Suite/Extended-Toolkit/issues/324), NuGet Installation Error in v65.22.6.152 Version - NU1103 Rollback